+++ src/pos_tui/widgets/__init__.py
@@ -5,6 +5,7 @@ This module contains custom widgets used throughout the POS application.
 """
 
 # Import widgets as they are implemented
+from .filter_bar import FilterBar
 from .item_form import ItemEntryForm
 from .item_table import ItemTable
 from .link_tree import LinkTree
@@ -14,6 +15,7 @@ from .link_tree import LinkTree
 __all__: list[str] = [
     "ItemEntryForm",
     "ItemTable",
+    "FilterBar",
     "LinkTree",
     # "Message",
 ]